refactor: move migrate_addon_settings to helpers
